Create add_files.sh
This file makes it easier to commit and push the >30k files of CK3 to the PRIVATE Github Repo of the Action. It will commit max 100 files at the same time and max 100 Mb at the same time. (I mainly added that because if i upload ~1Gb at the same time i always get timed out by my Internet Provider)